Fix id type mismatch: update Rust models and DB schema to i64
* Updated `sessions` and `users` models to use `i64` for `id` field.
* Added `i64_format` helper to serialize `id` as string (to avoid JS precision issues) and deserialize from string or number.
* Updated migrations for `sessions` and `users` to explicitly use `big_integer` (INT8) for `id` to ensure compatibility between Rust code and Database schema.
* Updated `sessions` controller to use `i64`.

+++ backend/src/models/_entities/sessions.rs
@@ -9,7 +9,8 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    #[serde(with = "crate::models::i64_format")]
+    pub id: i64,
     pub title: Option<String>,
     #[sea_orm(column_type = "Text", nullable)]
     pub content: Option<String>,
