+++ backend/src/models/_entities/users.rs
@@ -9,7 +9,8 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    #[serde(with = "crate::models::i64_format")]
+    pub id: i64,
     pub pid: Uuid,
     #[sea_orm(unique)]
     pub email: String,
